Reorganize compenents in directory
