+++ frontend/app/src/components/Modals/DeleteModal.vue
